requirements.txt
flask
pandas
PyMySQL
python-dotenv
requests
schedule
slack-bolt
slack-sdk
